requirements.txt
qiskit==1.0.0
qiskit-ibm-provider>=0.7.0
qiskit-ibm-runtime>=0.23.0
qiskit-machine-learning>=0.7.0
numpy>=1.21.0
fastapi>=0.70.0
uvicorn>=0.17.0
uuid>=1.0.0
pydantic>=1.8.0
python-jose
passlib[bcrypt]